requirements.txt
asgiref==3.3.4
click==8.0.1
Django==3.2.4
loguru==0.5.3
psycopg2==2.8.6
python-dotenv==0.17.1
pytz==2021.1
sqlparse==0.4.1
